Assertions in the tests now verify that libsshrb's C component is compiling correctly.
The c-module is called libsshrb, the ruby include simply `libssh`. This
seems a little uncomfortable, but will be easy to change in the future.
For users, all there is to know is that one must `require "libssh"`, and
then use methods in the `LibSSH` module.

I believe the two existing unit tests are sufficient to test the current
status, and I will begin work on the configuration class, and simple
tests for connection pooling, and so-forth.